Move most flags from module_codegen to new_context
This way they also apply to the allocator shim.
